Refactor: 폴더구조 변경(API 관련 코드는 /api 로 이동)
