+++ src/api/feed/feed.module.ts
@@ -1,7 +1,7 @@
 import {Module} from '@nestjs/common';
 import {FeedService} from './feed.service';
 import {FeedController} from './feed.controller';
-import {ImageModule} from 'src/image/image.module';
+import {ImageModule} from 'src/api/image/image.module';
 import {BlogChallenges} from 'src/entities/BlogChallenges';
 import {BlogImage} from 'src/entities/BlogImage';
 import {BlogPost} from 'src/entities/BlogPost';
