+++ src/api/image/image.controller.ts
