+++ src/api/image/image.module.ts
