+++ src/api/promotions/promotions.module.ts
