+++ src/app.module.ts
@@ -3,11 +3,11 @@ import {TypeOrmModule} from '@nestjs/typeorm';
 import {AppController} from './app.controller';
 import {AppService} from './app.service';
 import {AuthModule} from './auth/auth.module';
-import {ChallengesModule} from './challenges/challenges.module';
+import {ChallengesModule} from './api/challenges/challenges.module';
 import {ormConfig} from './config/orm';
-import {UsersModule} from './users/users.module';
-import {FeedModule} from './feed/feed.module';
-import {PromotionsModule} from './promotions/promotions.module';
+import {UsersModule} from './api/users/users.module';
+import {FeedModule} from './api/feed/feed.module';
+import {PromotionsModule} from './api/promotions/promotions.module';
 
 @Module({
   imports: [
