+++ src/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import {forwardRef, Module} from '@nestjs/common';
 import {JwtModule} from '@nestjs/jwt';
 import {PassportModule} from '@nestjs/passport';
-import {UsersModule} from '../users/users.module';
+import {UsersModule} from '../api/users/users.module';
 import {AuthService} from './auth.service';
 import {JwtStrategy} from './jwt/jwt.strategy';
